Merges Windows-specific build improvements:
- Move remaining dependencies to vcpkg where available
- Normalize path handling and local file resolution
- Work around MSVC warning noise
- Bump LLVM to 21.1.8
- Clean up CMake, CI, and installer configuration

No functional or cross-platform behavior changes.

+++ src/MayaFlux/Buffers/Container/SoundFileBridge.cpp
@@ -23,7 +23,7 @@ void SoundFileBridge::setup_processors(Buffers::ProcessingToken token)
 {
     m_capture_stream = std::make_shared<Kakshya::DynamicSoundStream>(
         Config::get_sample_rate(),
-        get_container()->get_structure().get_channel_count());
+        static_cast<uint32_t>(get_container()->get_structure().get_channel_count()));
 
     m_stream_writer = std::make_shared<SoundStreamWriter>(m_capture_stream);
